+++ packages/core/src/components/sections/ProductGallery/ProductGallery.tsx
@@ -68,6 +68,8 @@ function ProductGallerySection({
   )
 }
 
+ProductGallerySection.$componentKey = 'ProductGallery'
+
 const OverridableProductGallery = getOverridableSection<
   typeof ProductGallerySection
 >('ProductGallery', ProductGallerySection, ProductGalleryDefaultComponents)