+++ packages/core/src/pages/404.tsx
@@ -8,6 +8,7 @@ import {
 
 import { default as GLOBAL_COMPONENTS } from 'src/components/cms/global/Components'
 import RenderSections from 'src/components/cms/RenderSections'
+import { getComponentKey } from 'src/utils/cms'
 import { OverriddenDefaultEmptyState as EmptyState } from 'src/components/sections/EmptyState/OverriddenDefaultEmptyState'
 import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import PLUGINS_COMPONENTS from 'src/plugins'
@@ -20,7 +21,7 @@ import type { PreviewData } from 'src/server/content/types'
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
   ...GLOBAL_COMPONENTS,
-  EmptyState,
+  [getComponentKey(EmptyState, 'EmptyState')]: EmptyState,
   ...PLUGINS_COMPONENTS,
   ...CUSTOM_COMPONENTS,
 }